+++ build.gradle
@@ -4,40 +4,46 @@ plugins {
     id "io.ktor.plugin" version "2.2.3"
 }
 
-
-group 'edu.romoshi'
-version '1.0-VERSION'
-
-application {
-    mainClass.set("edu.romoshi.Main")
+subprojects {
+    apply plugin: 'java'
+    apply plugin: 'application'
+    apply plugin: 'io.ktor.plugin'
 }
 
-ktor {
-    fatJar {
-        archiveFileName.set("t-bot.jar")
+allprojects {
+    group 'edu.romoshi'
+    version '1.0-VERSION'
+
+    application {
+        mainClass.set("edu.romoshi.Main")
+    }
+
+    ktor {
+        fatJar {
+            archiveFileName.set("t-bot.jar")
+        }
+    }
+
+    repositories {
+        mavenCentral()
+        maven { url 'https://jitpack.io' }
+    }
+
+    dependencies {
+        implementation 'mysql:mysql-connector-java:8.0.31'
+        implementation group: 'commons-codec', name: 'commons-codec', version: '1.9'
+        implementation group: 'org.telegram', name: 'telegrambots', version: '6.4.0'
+        implementation group: 'at.favre.lib', name: 'bcrypt', version: '0.9.0'
+
+        implementation group: 'org.slf4j', name: 'slf4j-api', version: '2.0.6'
+        implementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.4.5'
+
+        testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
+        testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+        testImplementation group: 'org.mockito', name: 'mockito-core', version: '5.1.1'
+    }
+
+    test {
+        useJUnitPlatform()
     }
 }
-
-repositories {
-    mavenCentral()
-    maven { url 'https://jitpack.io' }
-}
-
-dependencies {
-    implementation 'mysql:mysql-connector-java:8.0.31'
-    implementation group: 'commons-codec', name: 'commons-codec', version: '1.9'
-    implementation group: 'org.telegram', name: 'telegrambots', version: '6.4.0'
-    implementation group: 'at.favre.lib', name: 'bcrypt', version: '0.9.0'
-
-    implementation group: 'org.slf4j', name: 'slf4j-api', version: '2.0.6'
-    implementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.4.5'
-
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    testImplementation group: 'org.mockito', name: 'mockito-core', version: '5.1.1'
-}
-
-test {
-    useJUnitPlatform()
-}
-
