+++ polling/build.gradle
@@ -1,19 +1,3 @@
-plugins {
-    id 'java'
-}
-
-group 'edu.romoshi'
-version '1.0-VERSION'
-
-repositories {
-    mavenCentral()
-}
-
 dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-}
-
-test {
-    useJUnitPlatform()
+    implementation project(":")
 }
