+++ polling/src/main/java/edu/romoshi/bot/BotStrings.java
