+++ polling/src/main/java/edu/romoshi/bot/commands/CommandStrings.java
