+++ polling/src/main/java/edu/romoshi/bot/commands/DefaultCommand.java
@@ -4,17 +4,17 @@ import edu.romoshi.bot.BotStrings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.telegram.telegrambots.meta.api.objects.Message;
+
 import static edu.romoshi.Main.bot;
 
-public class HelpCommand implements Command {
-    private static final Logger logger = LoggerFactory.getLogger(HelpCommand.class);
-
+public class DefaultCommand implements Command {
+    private static final Logger logger = LoggerFactory.getLogger(DefaultCommand.class);
     @Override
     public void execute(Message message) {
         try {
-            bot.sendMsg(message, BotStrings.HELP_STRING);
+            bot.sendMsg(message, BotStrings.DEFAULT_STRING);
         } catch (Exception ex) {
-            logger.error("Help command", ex);
+            logger.error("Def command", ex);
         }
     }
 }
