+++ polling/src/main/java/edu/romoshi/bot/commands/HelpCommand.java
@@ -1,6 +1,5 @@
 package edu.romoshi.bot.commands;
 
-import edu.romoshi.Main;
 import edu.romoshi.bot.BotStrings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,15 +7,15 @@ import org.telegram.telegrambots.meta.api.objects.Message;
 
 import static edu.romoshi.Main.bot;
 
-public class StartCommand implements Command {
-    private static final Logger logger = LoggerFactory.getLogger(StartCommand.class);
+public class HelpCommand implements Command {
+    private static final Logger logger = LoggerFactory.getLogger(HelpCommand.class);
 
     @Override
     public void execute(Message message) {
         try {
-            bot.sendMsg(message, BotStrings.START_STRING);
+            bot.sendMsg(message, BotStrings.HELP_STRING);
         } catch (Exception ex) {
-            logger.error("Start command", ex);
+            logger.error("Help command", ex);
         }
     }
 }
