+++ polling/src/main/java/edu/romoshi/bot/commands/KeyCommand.java
@@ -1,13 +1,14 @@
 package edu.romoshi.bot.commands;
 
 import at.favre.lib.crypto.bcrypt.BCrypt;
-import edu.romoshi.Main;
 import edu.romoshi.bot.BotStrings;
 import edu.romoshi.jdbc.users.Users;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.telegram.telegrambots.meta.api.objects.Message;
+
 import static edu.romoshi.Main.bot;
+
 
 public class KeyCommand implements Command {
     private static final Logger logger = LoggerFactory.getLogger(KeyCommand.class);
