+++ polling/src/main/java/edu/romoshi/bot/commands/ShowCommand.java
@@ -1,28 +1,32 @@
 package edu.romoshi.bot.commands;
 
-import edu.romoshi.Main;
 import edu.romoshi.bot.BotStrings;
 import edu.romoshi.jdbc.accounts.Accounts;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.telegram.telegrambots.meta.api.objects.Message;
+import java.util.List;
+
 import static edu.romoshi.Main.bot;
 
-public class DeleteCommand implements Command {
-    private static final Logger logger = LoggerFactory.getLogger(DeleteCommand.class);
+public class ShowCommand implements Command {
+    private static final Logger logger = LoggerFactory.getLogger(ShowCommand.class);
     private final boolean verifyKey;
 
-    public DeleteCommand(boolean verifyKey) {
+    public ShowCommand(boolean verifyKey) {
         this.verifyKey = verifyKey;
     }
+
     @Override
     public void execute(Message message) {
         String[] messageArray = message.getText().split(" ");
         try {
             if (verifyKey) {
-                if(messageArray.length == 2) {
-                    Accounts.deleteAccount(messageArray[1], message);
-                    bot.sendMsg(message, BotStrings.DELETE_STRING);
+                if(messageArray.length == 1) {
+                    List<Accounts> accounts =Accounts.getAccounts(message);
+                    for (var account : accounts) {
+                        bot.sendMsg(message, account.getInfo(message));
+                    }
                 } else {
                     bot.sendMsg(message, BotStrings.MISTAKE_MESSAGE);
                 }
@@ -30,7 +34,7 @@ public class DeleteCommand implements Command {
                 bot.sendMsg(message, BotStrings.START_STRING);
             }
         } catch (Exception ex) {
-            logger.error("Delete command", ex);
+            logger.error("Show command", ex);
         }
     }
 }
