+++ polling/src/main/java/edu/romoshi/bot/commands/StartCommand.java
@@ -1,21 +1,21 @@
 package edu.romoshi.bot.commands;
 
-
-import edu.romoshi.Main;
 import edu.romoshi.bot.BotStrings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.telegram.telegrambots.meta.api.objects.Message;
+
 import static edu.romoshi.Main.bot;
 
-public class DefaultCommand implements Command {
-    private static final Logger logger = LoggerFactory.getLogger(DefaultCommand.class);
+public class StartCommand implements Command {
+    private static final Logger logger = LoggerFactory.getLogger(StartCommand.class);
+
     @Override
     public void execute(Message message) {
         try {
-            bot.sendMsg(message, BotStrings.DEFAULT_STRING);
+            bot.sendMsg(message, BotStrings.START_STRING);
         } catch (Exception ex) {
-            logger.error("Def command", ex);
+            logger.error("Start command", ex);
         }
     }
 }
