--- src/main/java/edu/romoshi/Main.java
@@ -1,25 +1,0 @@
-package edu.romoshi;
-
-import edu.romoshi.bot.Bot;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-
-public class Main {
-    private static final Logger logger = LoggerFactory.getLogger(Main.class);
-    public static final Bot bot = new Bot();
-
-    public static void main(String[] args) {
-        try {
-            TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
-            botsApi.registerBot(bot);
-            logger.info("Bot ON");
-        } catch (TelegramApiException e) {
-            logger.error("Problems with bot register", e);
-        }
-    }
-}
